requirements.txt
Flask
beautifulsoup4
python-docx